+++ connecteur-type/utilities/GetModuleList.php
@@ -1,5 +1,7 @@
 <?php
 
+use Pastell\Service\Module\ModuleListService;
+
 class GetModuleList extends ConnecteurTypeChoiceActionExecutor
 {
     private const MODULE_TYPE_FIELD = 'module_type';
@@ -10,9 +12,9 @@ class GetModuleList extends ConnecteurTypeChoiceActionExecutor
      * @return bool
      * @throws RecoverableException
      */
-    public function go()
+    public function go(): bool
     {
-        $moduleType = $this->getRecuperateur()->get(self::MODULE_TYPE_FIELD);
+        $moduleType = (string)$this->getRecuperateur()->get(self::MODULE_TYPE_FIELD);
         $moduleList = $this->displayAPI();
         if ($moduleType && empty($moduleList[$moduleType])) {
             throw new RecoverableException("Ce type de dossier n'existe pas");
@@ -31,22 +33,14 @@ class GetModuleList extends ConnecteurTypeChoiceActionExecutor
     /**
      * @throws NotFoundException
      */
-    public function display()
+    public function display(): true
     {
         $this->setViewParameter(
             'moduleType',
             $this->getConnecteurProperties()->get($this->getMappingValue(self::MODULE_TYPE_FIELD))
         );
-        $modules = $this->displayAPI();
 
-        $currentLocale = setlocale(LC_COLLATE, '0');
-        setlocale(LC_COLLATE, 'fr_FR.utf8');
-        uasort($modules, static function (array $a, array $b) {
-            return strcoll($a['nom'], $b['nom']);
-        });
-        setlocale(LC_COLLATE, $currentLocale);
-
-        $this->setViewParameter('moduleList', $modules);
+        $this->setViewParameter('moduleList', $this->displayAPI());
         $this->renderPage(
             $this->getMappingValue(self::PAGE_TITLE),
             'connectorType/utilities/GetModuleList'
@@ -54,8 +48,8 @@ class GetModuleList extends ConnecteurTypeChoiceActionExecutor
         return true;
     }
 
-    public function displayAPI()
+    public function displayAPI(): array
     {
-        return $this->apiGet("/flux", []);
+        return $this->objectInstancier->getInstance(ModuleListService::class)->getModuleListOrderByNom($this->id_u);
     }
 }