+++ CHANGELOG.md
@@ -97,6 +97,10 @@ doivent être mises à jour #1667
 - Suppression de ActionChange::updateModification, use addOrUpdateAction instead #2336
 - Suppression d'éléments dépréciés de util.php 
   (rrmdir, get_argv, exceptionToJson, throwIfFalse, throwLastError, wl_basename, tick, utf8_decode_array) #2347
+- Suppression de BaseAPIControllerFactory #2351
+- Suppression de BaseAPIController::hasOneDroit, FluxAPIController::listFlux, PastellController::getAPIController,
+  Gabarit::getAPIController, ActionExecutor::apiGet, ActionExecutor::apiCall, DocumentTypeFactory::isSuperTypePresent,
+  PastellTestCase::getAPIController #2351
 
 ## Dépréciations